add a common header file for string_t based maps
